requirements.txt
python3-magic
python3-yara
libnotify-bin
pulseaudio-utils
yara
libyara-dev
clamav
inotify-tools